basis functionaliteit

+++ MC1000/Controllers/HomeController.cs
@@ -55,7 +55,7 @@ namespace MC1000.Controllers
                         FullDescription = product.Descendants("Fulldescription").First().Value,
                         Image = product.Descendants("Image").First().Value,
                         Weight = product.Descendants("Weight").First().Value,
-                        Price = Decimal.Parse(product.Descendants("Price").First().Value, style, provider)
+                        Price = double.Parse(product.Descendants("Price").First().Value, style, provider)
                     };
                     _context.Add(p);
                 }
